[css-ui] Use float for outline-offset
It's expected that outline-offset accepts subpixel values,
the parsing actually accepts them, but then we store things as ints
so we lose the decimal precision.

This patch changes outline-offset to use a float to fix this issue.

Rebaselined some tests that use "zoom=1.25" and now have
a decimal outline-offset for real.

BUG=1068107
TEST=external/wpt/css/css-ui/outline-offset-002.html
TEST=external/wpt/css/css-ui/parsing/outline-offset-computed.html

Change-Id: I3ed0495150d37f9c14f357f38eb2787f00da5c08